WIP Retract existing singular attr value based on data, not delta
Somehow it seems that we cannot rely on the form delta to have the correct :before (though the error could also be elsewhere). I ended up with multiple values for an attribute that should be singular. With this refactoring we find out and retract the actual db value of the attribute with transacting the new one.