Retract existing singular attr value based on data, not delta (#2)
Somehow it seems that we cannot rely on the form delta to have the correct :before (though the error could also be elsewhere). I ended up with multiple values for an attribute that should be singular. With this refactoring we find out and retract the actual db value of the attribute with transacting the new one.

CORRECTION: The multiple vals of singular attributes were most likely caused by incorrect, repeated seeding of the db, not by any issue with how save-form! and its retractions were implemented.

Also

* reset db before each test
* small improvements, docs