Owned entities (#4)
Add support for `::asami/owned-entity?` on an ID attribute and auto-deleting such entity when removed from the parent's attribute that refers to it.

This makes it much easier to implement patterns such as an OrderLine, that can only exist as a part of an Order. When you remove it from the parent entity, Fulcro will only send a diff for the parent's attribute change - but the adapter will pick it up and also delete the whole owned entity. (Otherwise we'd need to do some trickery in the save middleware - but since we need to delete each and every inidividual attribute associated with the entity, the best place for that is as close to the DB as possibly, and thus this change.)

Also included: refactoring, linting cleanup, adding support for building a JAR of the library.